mark etl.TAEWriter and etl.TAEReader Deprecated.
